Edit suggested by Sullivan Marafico
Dear Leonel,

Sorry for not being clear, wgen is here for historical reason and plays no role when creating the tensor. I would recommand you to ignore it and change line 93 of LoadInput.C
From vwd.push_back(wd*dzdt); to vwd.push_back(1); for now.

In the future, we should re-write this part and get rid of this wgen once and for all.

Cheers,

Sullivan Marafico,
PhD Student IJCLAB